# Conflicts:
#	yggdrash-core/src/main/java/io/yggdrash/contract/CoinContract.java
#	yggdrash-core/src/main/java/io/yggdrash/core/NodeManager.java
#	yggdrash-core/src/main/java/io/yggdrash/core/Runtime.java
#	yggdrash-core/src/main/java/io/yggdrash/core/TransactionHusk.java
#	yggdrash-core/src/test/java/io/yggdrash/core/contract/CoinContractTest.java
#	yggdrash-node/src/main/java/io/yggdrash/node/NodeManagerImpl.java
#	yggdrash-node/src/main/java/io/yggdrash/node/api/AccountApi.java
#	yggdrash-node/src/main/java/io/yggdrash/node/api/AccountApiImpl.java
#	yggdrash-node/src/main/java/io/yggdrash/node/api/TransactionApiImpl.java

+++ yggdrash-node/src/main/java/io/yggdrash/node/api/AccountApi.java
@@ -47,7 +47,7 @@ public interface AccountApi {
     @JsonRpcErrors({
             @JsonRpcError(exception = NonExistObjectException.class,
                     code = NonExistObjectException.code)})
-    int getBalance(@JsonRpcParam(value = "address") String address,
+    long getBalance(@JsonRpcParam(value = "address") String address,
                    @JsonRpcParam(value = "blockNumber") int blockNumber);
 
     /**
@@ -59,6 +59,6 @@ public interface AccountApi {
     @JsonRpcErrors({
             @JsonRpcError(exception = NonExistObjectException.class,
                     code = NonExistObjectException.code)})
-    int getBalance(@JsonRpcParam(value = "address") String address,
+    long getBalance(@JsonRpcParam(value = "address") String address,
                    @JsonRpcParam(value = "tag") String tag);
 }
