+++ yggdrash-node/src/main/java/io/yggdrash/node/api/AccountApiImpl.java
@@ -18,7 +18,7 @@ import java.util.ArrayList;
 public class AccountApiImpl implements AccountApi {
 
     private final ArrayList<String> addresses = new ArrayList<>();
-    private final int balance = 100000;
+    private final long balance = 100000;
     private final NodeManager nodeManager;
     private final Runtime runtime;
     private CoinContract coinContract;
@@ -65,7 +65,7 @@ public class AccountApiImpl implements AccountApi {
     }
 
     @Override
-    public int getBalance(String address, int blockNumber) {
+    public long getBalance(String address, int blockNumber) {
         try {
             return balance;
         } catch (Exception exception) {
@@ -74,7 +74,7 @@ public class AccountApiImpl implements AccountApi {
     }
 
     @Override
-    public int getBalance(String address, String tag) {
+    public long getBalance(String address, String tag) {
         try {
             return balance;
         } catch (Exception exception) {
